requirements.txt
torch
torchvision==0.15.2 -f https://download.pytorch.org/whl/torch_stable.html
numpy==1.25.2
scipy==1.11.2
opencv-python
gdown
pillow==10.0.0
gradio